sanitize and slugify usernames; share some types;
used slugify package since it's used in the client too..
