+++ client/src/ui/loadouts.ts
@@ -1,96 +1,12 @@
 import { GameObjectDefs } from "../../../shared/defs/gameObjectDefs";
 import type { UnlockDefs } from "../../../shared/defs/gameObjects/unlockDefs";
-import { GameConfig } from "../../../shared/gameConfig";
-import type { Crosshair } from "../crosshair";
+import { ItemStatus, type Loadout, validateLoadout } from "../../../shared/utils/helpers";
 import { deepEqual } from "../lib/deepEqual";
 import type { Item } from "./loadoutMenu";
 
-export interface Loadout {
-    player_icon: string;
-    outfit: string;
-    heal: string;
-    boost: string;
-    melee: string;
-    emotes: string[];
-    crosshair: Crosshair;
-}
-
-export enum ItemStatus {
-    New,
-    Confirmed,
-    Ackd,
-}
-
 const loadout = {
     ItemStatus,
-    validate: function (userLoadout: Loadout) {
-        const getGameType = function (
-            type: string,
-            gameType: string,
-            defaultValue: string,
-        ) {
-            const def = GameObjectDefs[gameType];
-            if (def && def.type == type) {
-                return gameType;
-            }
-            return defaultValue;
-        };
-        const getFloat = function (flt: string, defaultValue: number) {
-            const val = parseFloat(flt);
-            if (Number.isNaN(val)) {
-                return defaultValue;
-            }
-            return val;
-        };
-        const mergedLoadout = {
-            ...{
-                crosshair: {
-                    type: "",
-                    color: 0xffffff,
-                    size: 1,
-                    stroke: 0,
-                },
-                emotes: [],
-            },
-            ...userLoadout,
-        } as Loadout;
-        const validatedLoadout: Loadout = {
-            outfit: getGameType("outfit", mergedLoadout.outfit, "outfitBase"),
-            melee: getGameType("melee", mergedLoadout.melee, "fists"),
-            heal: getGameType("heal_effect", mergedLoadout.heal, "heal_basic"),
-            boost: getGameType("boost_effect", mergedLoadout.boost, "boost_basic"),
-            player_icon: getGameType("emote", mergedLoadout.player_icon, ""),
-            crosshair: {
-                type: getGameType(
-                    "crosshair",
-                    mergedLoadout.crosshair.type,
-                    "crosshair_default",
-                ),
-                color:
-                    parseInt(mergedLoadout.crosshair.color as unknown as string) ||
-                    0xffffff,
-                size: getFloat(
-                    mergedLoadout.crosshair.size as unknown as string,
-                    1,
-                ).toFixed(2) as unknown as number,
-                stroke: getFloat(
-                    mergedLoadout.crosshair.stroke as unknown as string,
-                    0,
-                ).toFixed(2) as unknown as number,
-            },
-            emotes: [] as string[],
-        };
-
-        const defaultEmotes = GameConfig.defaultEmoteLoadout.slice();
-        for (let i = 0; i < GameConfig.EmoteSlot.Count; i++) {
-            const inputEmote =
-                i < mergedLoadout.emotes.length ? mergedLoadout.emotes[i] : "";
-            validatedLoadout.emotes.push(
-                getGameType("emote", inputEmote, defaultEmotes[i]),
-            );
-        }
-        return validatedLoadout;
-    },
+    validate: validateLoadout,
     /* not used
     validateWithAvailableItems: function(userLoadout, userItems) {
         const checkTypeExists = function(type, items) {
