+++ server/package.json
@@ -20,6 +20,7 @@
     "hono": "^4.5.9",
     "lucia": "^3.2.0",
     "nanotimer": "^0.3.15",
+    "slugify": "^1.6.6",
     "uWebSockets.js": "github:uNetworking/uWebSockets.js#v20.48.0",
     "zod": "^3.23.8"
   },
