+++ server/src/api/auth/middleware.ts
@@ -1,5 +1,5 @@
 import type { Context } from "hono";
-import { deleteCookie, getCookie, setCookie } from "hono/cookie";
+import { deleteCookie, getCookie } from "hono/cookie";
 import { lucia } from "./lucia";
 
 export const AuthMiddleware = async (c: Context, next: () => Promise<void>) => {
