+++ server/src/api/db/drizzle/meta/_journal.json
@@ -5,8 +5,8 @@
     {
       "idx": 0,
       "version": "6",
-      "when": 1726315233497,
-      "tag": "0000_abnormal_dormammu",
+      "when": 1726827855683,
+      "tag": "0000_gigantic_brother_voodoo",
       "breakpoints": true
     }
   ]
